+++ structures/dynamic_hash_table_iterator.h
@@ -6,7 +6,7 @@
 
 struct bsal_dynamic_hash_table_iterator {
     struct bsal_dynamic_hash_table *list;
-    int index;
+    uint64_t index;
 };
 
 void bsal_dynamic_hash_table_iterator_init(struct bsal_dynamic_hash_table_iterator *self, struct bsal_dynamic_hash_table *list);
